+++ samples/modules/pmci/index.rst
@@ -1,0 +1,5 @@
+.. zephyr:code-sample-category:: pmci
+   :name: PMCI
+   :show-listing:
+
+   These samples demonstrate the use of PMCI related protocols in Zephyr.
