+++ samples/modules/pmci/mctp/endpoint/CMakeLists.txt
