+++ samples/modules/pmci/mctp/endpoint/README.rst
@@ -29,7 +29,7 @@ Building and Running
 
 
 .. zephyr-app-commands::
-   :zephyr-app: samples/modules/mctp/mctp_endpoint
+   :zephyr-app: samples/modules/pmci/mctp/endpoint
    :host-os: unix
    :board: nrf52840_nrf52840dk
    :goals: run
