+++ samples/modules/pmci/mctp/endpoint/boards/nrf52840dk_nrf52840.overlay
