+++ samples/modules/pmci/mctp/endpoint/prj.conf
