+++ samples/modules/pmci/mctp/host/boards/nrf52840dk_nrf52840.overlay
