+++ subsys/CMakeLists.txt
@@ -22,6 +22,7 @@ add_subdirectory(mem_mgmt)
 add_subdirectory(mgmt)
 add_subdirectory(modbus)
 add_subdirectory(pm)
+add_subdirectory(pmci)
 add_subdirectory(portability)
 add_subdirectory(random)
 add_subdirectory(rtio)
@@ -46,7 +47,6 @@ add_subdirectory_ifdef(CONFIG_IMG_MANAGER dfu)
 add_subdirectory_ifdef(CONFIG_INPUT input)
 add_subdirectory_ifdef(CONFIG_JWT jwt)
 add_subdirectory_ifdef(CONFIG_LLEXT llext)
-add_subdirectory_ifdef(CONFIG_MCTP mctp)
 add_subdirectory_ifdef(CONFIG_MODEM_MODULES modem)
 add_subdirectory_ifdef(CONFIG_NETWORKING net)
 add_subdirectory_ifdef(CONFIG_PROFILING profiling)
