+++ subsys/Kconfig
@@ -26,13 +26,13 @@ source "subsys/jwt/Kconfig"
 source "subsys/llext/Kconfig"
 source "subsys/logging/Kconfig"
 source "subsys/lorawan/Kconfig"
-source "subsys/mctp/Kconfig"
 source "subsys/mem_mgmt/Kconfig"
 source "subsys/mgmt/Kconfig"
 source "subsys/modbus/Kconfig"
 source "subsys/modem/Kconfig"
 source "subsys/net/Kconfig"
 source "subsys/pm/Kconfig"
+source "subsys/pmci/Kconfig"
 source "subsys/portability/Kconfig"
 source "subsys/profiling/Kconfig"
 source "subsys/random/Kconfig"
