+++ subsys/pmci/CMakeLists.txt
@@ -1,0 +1,3 @@
+# SPDX-License-Identifier: Apache-2.0
+
+add_subdirectory(mctp)
