+++ subsys/pmci/mctp/mctp_uart.c
@@ -8,7 +8,7 @@
 #include <zephyr/sys/__assert.h>
 #include <zephyr/kernel.h>
 #include <zephyr/drivers/uart.h>
-#include <zephyr/mctp/mctp_uart.h>
+#include <zephyr/pmci/mctp/mctp_uart.h>
 #include <crc-16-ccitt.h>
 
 #include <zephyr/logging/log.h>
